requirements.txt
flask
python-dotenv
pycryptodomex 